requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pdfrw==0.4
pypdf==5.5.0
requests==2.32.3
urllib3==2.4.0
Werkzeug==3.1.3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

